+++ sphinx/source/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 9.6b1
 -----
+* Added the ``--press`` option to :ref:`tap2sna.py` (for pausing the tape and
+  :ref:`simulating keypresses <tap2sna-user-input>` before resuming)
 * Changed the default value of the ``accelerate-dec-a`` simulated LOAD
   configuration parameter for :ref:`tap2sna.py <tap2sna-sim-load>` from ``1``
   to ``3``